--- src/components/FormSign/index.js
@@ -1,1 +1,0 @@
-export * from './FormSign'
