+++ src/components/index.js
@@ -5,7 +5,7 @@ export * from './CardContainer'
 export * from './FormLogin'
 export * from './FieldPassword'
 export * from './DivError'
-export * from './FormSign'
+export * from './FormSignIn'
 export * from './DivError'
 export * from './FormTitle'
 export * from './FieldCpf'
